pyproject.toml
[project]
name = "ols-mcp"
version = "0.1.0"
description = "OpenShift LightSpeed MCP Server - Provides access to OpenShift LightSpeed service via MCP protocol"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mcp>=1.2.1",
    "pydantic>=2.0.0",
    "httpx>=0.25.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.10.0",
]
